Refactor: Separate component in Login page
