+++ src/pages/Login/components/LoginTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginTitle';
